:construction: FIrst upload.
This commit adds much of the original code of the `FFTFIT` algorithm
from `PRESTO`, and makes it installable as a Python package. We are
using a soon-to-be-deprecated `numpy` interface, but for now it works.
Still have to write some of the Python code (again adapted from
`PRESTO`), have to figure out some stuff about the licensing of the
code, and have to try and compile cross-platform wheels for PyPI (via
`cibuildwheel`).